Commit again and again.
