+++ src/main/java/com/antra/phonepad/combination/Controller/CombinationController.java
@@ -1,9 +1,7 @@
 package com.antra.phonepad.combination.Controller;
 
-import java.io.IOException;
 import java.util.List;
 
-import com.antra.phonepad.combination.service.ToNumMapperService;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.beans.factory.annotation.Autowired;
 
@@ -12,17 +10,17 @@ import com.antra.phonepad.combination.service.ToLetterMapperService;
 @RestController
 @RequestMapping()
 public class CombinationController {
+
     @Autowired
     ToLetterMapperService toLetterMapperService;
-    @Autowired
-    ToNumMapperService toNumMapperService;
+
     @RequestMapping(value = "/toNumbers/{num}", method = RequestMethod.GET)
     public String toNumbers(@PathVariable("num") String num) {
        if (num.length() != 10) {
             throw new ArithmeticException("Invalid telephone numbers. Must be 10 digits");
        } else {
             num = num.toUpperCase();
-            return toNumMapperService.ToNumMapper(num);
+            return toLetterMapperService.ToNumMapper(num);
        }
     }
