--- src/main/java/com/antra/phonepad/combination/dao/ToNumMapper.java
@@ -1,5 +1,0 @@
-package com.antra.phonepad.combination.dao;
-
-public interface ToNumMapper {
-    String toNumMapper(String num);
-}
