+++ src/main/java/com/antra/phonepad/combination/service/ToLetterMapperService.java
@@ -3,4 +3,6 @@ import java.util.List;
 
 public interface ToLetterMapperService {
     List<String> ToLetterMapper(String num);
+
+    String ToNumMapper (String num);
 }
