--- src/main/java/com/antra/phonepad/combination/service/ToNumMapperService.java
@@ -1,6 +1,0 @@
-package com.antra.phonepad.combination.service;
-
-
-public interface ToNumMapperService {
-    String ToNumMapper (String num);
-}
